Project/requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
fpdf>=1.7.2
Werkzeug>=2.0.0
